修复删除目录方法，会将软连接指向的目录也删除的问题

+++ src/Util/File.php
@@ -262,7 +262,7 @@ class File
             if ('.' !== $file && '..' !== $file)
             {
                 $fullpath = $dir . '/' . $file;
-                if (is_dir($fullpath))
+                if (is_dir($fullpath) && !is_link($fullpath))
                 {
                     self::deleteDir($fullpath);
                 }
